style(lib): move unrelated methods into lib
circleci.ts  no longer had a reason for holding constants and helpers
